requirements.txt
aenum==3.1.15
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.11.0
asttokens==3.0.0
attrs==24.3.0
azure-core==1.32.0
azure-storage-blob==12.24.0
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.0
case_conversion==2.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
clipboard==0.0.4
colorama==0.4.6
colour==0.1.5
comm==0.2.2
crashtest==0.4.1
cryptography==44.0.0
cursor==1.3.5
debugpy==1.8.11
decorator==5.1.1
dill==0.3.9
distlib==0.3.9
dulwich==0.22.7
et_xmlfile==2.0.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fuzzywuzzy==0.18.0
gender-guesser==0.4.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==5.12.4
gspread-pandas==3.3.0
html_text==0.6.2
httplib2==0.22.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
infi.systray==0.1.12
inflect==7.5.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.29.5
ipython==8.31.0
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
lxml==5.3.0
lxml_html_clean==0.4.1
matplotlib-inline==0.1.7
maybe-else==0.2.1
mbstrdecoder==1.1.3
mock==5.1.0
more-itertools==10.5.0
moviepy==2.1.1
msal==1.31.1
msgpack==1.1.0
msoffcrypto-tool==5.4.2
multipledispatch==1.0.0
natsort==8.4.0
nest-asyncio==1.6.0
numpy==2.2.1
O365==2.0.38
oauthlib==3.2.2
office365==0.3.15
Office365-REST-Python-Client==2.5.14
olefile==0.47
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandas-flavor==0.6.0
parsedatetime==2.6
parso==0.8.4
pathmagic==0.3.14
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
prettierfier==1.0.3
proglog==0.1.10
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
PureCloudPlatformClientV2==219.1.0
-e git+ssh://git@github.com/ChrisPachulski/py_toolkit.git@cb31c82cce473de7829fdc0493b433c1a056732c#egg=py_toolkit
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
pyinstrument==5.0.0
pyiotools==0.3.18
pyjanitor==0.30.0
PyJWT==2.10.1
pymiscutils==0.3.14
pyparsing==3.2.1
PyPDF2==3.0.1
pyperclip==1.9.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5
PyQt5_sip==12.16.1
pysubtypes==0.3.18
pytest==8.3.4
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
RapidFuzz==3.11.0
readchar==4.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
scipy==1.15.1
Send2Trash==1.8.3
setuptools==75.8.0
shellingham==1.5.4
simple-salesforce==1.12.6
simplejson==3.19.3
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tabulate==0.9.0
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.1.15.22
typeguard==4.4.1
typepy==1.3.4
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.1
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
xarray==2025.1.1
xattr==1.1.4
XlsxWriter==3.2.0
zeep==4.3.1
